Add user signup confirm functionality
Added a confirmation procedure to the user signup flow. This includes a new Command, Handler and tests for the confirmation process. Also, some changes were made in the User entity and UserRepository to support this functionality. An extra method "confirmSignUp" was added to the User entity to validate the confirmation process.