Refactor User class and update corresponding test files
Transformed the User class to utilize static creation methods for signing up users. Rewritten corresponding entity and user test files to be compatible with the changes. Removed the CreateTest.php file, as the change in implementation rendered it obsolete. Added port mapping for the database service in the compose.yaml file.