Refactor User and Network entity annotations
Revised the annotations in User and Network entities. The refactoring included the removal of explicit string type declaration for 'passwordHash' and 'confirmToken' properties in User. Additionally, ORM mappings for 'networks' were established. In the Network entity, ORM annotations were added to define it as an entity while detailing column types, relationships, and constraints.