Add OAuth2 Facebook support and update configuration
This commit includes the addition of the OAuth2 client bundle and the OAuth2 Facebook provider to the project. Changes have been made to composer.json, and symfony.lock to amend the dependencies required for the project. A configuration file (knpu_oauth2_client.yaml) has also been created to set up the Facebook client for OAuth2. Changes in the codebase demand an update in the composer.lock file to reflect the newly added dependency state.