Add change role console command
Refactored the handling of user roles to better check if current role is null before changing. The role constants in the Role model were made public. In addition, a new RoleCommand was introduced to change user roles via console. Also implemented validation for not blank fields in the role changing use case.